feature: Make general simulate function
feature: Make general simulate function. Function can simulate either an equilibrium run or a run where the box is expanded and the atoms condense into a nanoparticle. As a part of this, input files are removed as they are highly similar, with only LJ parameters, atomic masses, lattice parameters, and crystal structure being different.